Moved the tira RestClient and LocalClient into an _internal package
