+++ python-client/tests/pd_load_data_test.py
@@ -1,9 +1,9 @@
-from tira.rest_api_client import Client
+from tira.tira_client import RestClient
 import unittest
 
 class TestLocalExecutionIntegrationTest(unittest.TestCase):
     def test_pd_load_inputs_wows_re_ranking(self):
-        tira = Client()
+        tira = RestClient()
         actual = tira.pd.inputs('workshop-on-open-web-search', 're-ranking-20231027-training')
         self.assertEqual(6, len(actual))
         self.assertEqual('hubble telescope achievements', actual.iloc[0]['query'])
@@ -12,7 +12,7 @@ class TestLocalExecutionIntegrationTest(unittest.TestCase):
         self.assertEqual('1', actual.iloc[0]['qid'])
 
     def test_pd_load_truths_wows_re_ranking(self):
-        tira = Client()
+        tira = RestClient()
         actual = tira.pd.truths('workshop-on-open-web-search', 're-ranking-20231027-training')
         self.assertEqual(3, len(actual))
         self.assertEqual('hubble telescope achievements', actual.iloc[0]['query'])
