+++ python-client/tests/pt_query_processing_loader_test.py
@@ -1,11 +1,11 @@
-from tira.local_client import Client
+from tira.tira_client import LocalClient, RestClient
 
 import pandas as pd
 import unittest
 
 def query_segmentation(queries):
     queries = pd.DataFrame([{'qid': str(i)} for i in queries])
-    tira = Client('tests/resources/')
+    tira = LocalClient('tests/resources/')
     query_segmentation =  tira.pt.transform_queries('ir-benchmarks/webis-query-segmentation/wt-snp-baseline', dataset='d1')
     
     ret = query_segmentation(queries)
@@ -34,27 +34,24 @@ class TestQueryProcessingLoader(unittest.TestCase):
         ensure_pyterrier_is_loaded()
 
     def test_rest_query_submission_with_rest_api_01(self):
-        from tira.rest_api_client import Client
-        tira = Client()
+        tira = RestClient()
 
         q = tira.pt.transform_queries('ir-benchmarks/ows/query-segmentation-wt-snp', 'disks45-nocr-trec-robust-2004-20230209-training')
         self.assertEqual(len(q(pd.DataFrame([{'qid': '303'}]))), 1)
         self.assertEqual(q(pd.DataFrame([{'qid': '303'}])).iloc[0].to_dict()['segmentation'], ['hubble telescope achievements'])
 
     def test_rest_query_submission_with_rest_api_02(self):
-        from tira.rest_api_client import Client
-        tira = Client()
+        tira = RestClient()
 
         q = tira.pt.transform_queries('ir-benchmarks/ows/query-segmentation-hyb-a', 'disks45-nocr-trec-robust-2004-20230209-training')
         self.assertEqual(len(q(pd.DataFrame([{'qid': '303'}]))), 1)
         self.assertEqual(q(pd.DataFrame([{'qid': '303'}])).iloc[0].to_dict()['segmentation'], ['hubble telescope', 'achievements'])
 
     def test_rest_query_submission_with_rest_api_03(self):
-        from tira.rest_api_client import Client
         from tira.third_party_integrations import ensure_pyterrier_is_loaded
         import pyterrier as pt
         ensure_pyterrier_is_loaded()
-        tira = Client()
+        tira = RestClient()
 
         q = tira.pt.transform_queries('ir-benchmarks/fschlatt/query-health-classification', pt.get_dataset("irds:clueweb09/en/trec-web-2009"))
         self.assertEqual(len(q(pd.DataFrame([{'qid': '26'}]))), 1)
@@ -62,11 +59,10 @@ class TestQueryProcessingLoader(unittest.TestCase):
         self.assertEqual(q(pd.DataFrame([{'qid': '26'}])).iloc[0].to_dict()['median_health_score'], 130.3137)
 
     def test_rest_query_submission_with_rest_api_and_custom_selection(self):
-        from tira.rest_api_client import Client
         from tira.third_party_integrations import ensure_pyterrier_is_loaded
         import pyterrier as pt
         ensure_pyterrier_is_loaded()
-        tira = Client()
+        tira = RestClient()
 
         q = tira.pt.transform_queries('ir-benchmarks/fschlatt/query-health-classification', pt.get_dataset("irds:clueweb09/en/trec-web-2009"), '/q*.jsonl')
         self.assertEqual(len(q(pd.DataFrame([{'qid': '26'}]))), 1)
@@ -74,11 +70,10 @@ class TestQueryProcessingLoader(unittest.TestCase):
         self.assertEqual(q(pd.DataFrame([{'qid': '26'}])).iloc[0].to_dict()['median_health_score'], 130.3137)
 
     def test_rest_query_submission_with_rest_api_04(self):
-        from tira.rest_api_client import Client
         from tira.third_party_integrations import ensure_pyterrier_is_loaded
         import pyterrier as pt
         ensure_pyterrier_is_loaded()
-        tira = Client()
+        tira = RestClient()
 
         q = tira.pt.transform_queries('ir-benchmarks/qpptk/fixed-sealer', pt.get_dataset("irds:disks45/nocr/trec-robust-2004"))
         self.assertEqual(len(q(pd.DataFrame([{'qid': '301'}]))), 1)
@@ -86,11 +81,10 @@ class TestQueryProcessingLoader(unittest.TestCase):
         self.assertEqual(q(pd.DataFrame([{'qid': '301'}])).iloc[0].to_dict()['avg-idf'], 2.4740487603)
 
     def test_rest_query_features(self):
-        from tira.rest_api_client import Client
         from tira.third_party_integrations import ensure_pyterrier_is_loaded
         import pyterrier as pt
         ensure_pyterrier_is_loaded()
-        tira = Client()
+        tira = RestClient()
 
         q = tira.pt.query_features('ir-benchmarks/qpptk/fixed-sealer', pt.get_dataset("irds:disks45/nocr/trec-robust-2004"))
         self.assertEqual(len(q(pd.DataFrame([{'qid': '301'}]))), 1)
