+++ python-client/tests/test_utils.py
@@ -1,7 +1,7 @@
 import requests
 from hashlib import md5
-from tira.tira_redirects import redirects, mirror_url
-from tira.rest_api_client import Client
+from tira.tira_redirects import redirects
+from tira.tira_client import RestClient
 
 def md5_of_first_kilobyte_of_http_resource(url):
     #if not url.startswith('https://files.webis.de'):
@@ -36,7 +36,7 @@ def digest_of_dataset(dataset_id, truth=False):
     }
 
 def digest_of_run_output(approach, dataset_id, run_ids):
-    tira = Client()
+    tira = RestClient()
     task, team, system = approach.split('/')
     run_url = f'https://www.tira.io/task/{task}/user/{team}/dataset/{dataset_id}/download/{run_ids[task][team][system][dataset_id]}.zip'
